fix: GPlayInEditorID가 비정상적으로 출력되던 문제를 해결

+++ Plugins/NexusAction/Source/NexusAction/Public/NexusLog.h
@@ -11,8 +11,8 @@
 #define LOG_REMOTEROLEINFO *(UEnum::GetValueAsString(TEXT("Engine.ENetRole"), GetRemoteRole()))
 #define LOG_REMOTEROLEINFO_SUB *(UEnum::GetValueAsString(TEXT("Engine.ENetRole"), Cast<AActor>(GetOuter())->GetRemoteRole()))
 
-#define LOG_NETMODEINFO ((GetNetMode() == ENetMode::NM_Client) ? *FString::Printf(TEXT("CLIENT%d"), GPlayInEditorID) : ((GetNetMode() == ENetMode::NM_Standalone) ? TEXT("STANDALONE") : TEXT("SERVER")))
-#define LOG_NETMODEINFO_SUB ((Cast<AActor>(GetOuter())->GetNetMode() == ENetMode::NM_Client) ? *FString::Printf(TEXT("CLIENT%d"), GPlayInEditorID) : ((Cast<AActor>(GetOuter())->GetNetMode() == ENetMode::NM_Standalone) ? TEXT("STANDALONE") : TEXT("SERVER")))
+#define LOG_NETMODEINFO ((GetNetMode() == ENetMode::NM_Client) ? *FString::Printf(TEXT("CLIENT%d"),  static_cast<int32>(GPlayInEditorID)) : ((GetNetMode() == ENetMode::NM_Standalone) ? TEXT("STANDALONE") : TEXT("SERVER")))
+#define LOG_NETMODEINFO_SUB ((Cast<AActor>(GetOuter())->GetNetMode() == ENetMode::NM_Client) ? *FString::Printf(TEXT("CLIENT%d"), static_cast<int32>(GPlayInEditorID)) : ((Cast<AActor>(GetOuter())->GetNetMode() == ENetMode::NM_Standalone) ? TEXT("STANDALONE") : TEXT("SERVER")))
 
 #define LOG_CALL_INFO ANSI_TO_TCHAR(__FUNCTION__)
 #define NX_LOG_FN(LogCat, Verbosity, Format, ...) UE_LOG(LogCat, Verbosity, TEXT("[%s][%s/%s] %s %s"), LOG_NETMODEINFO, LOG_LOCALROLEINFO, LOG_REMOTEROLEINFO, LOG_CALL_INFO, *FString::Printf(Format, ##__VA_ARGS__))
